Update 2 September 2019

+++ src/inventory/DlgInputStokPasien.java
@@ -55,6 +55,7 @@ public class DlgInputStokPasien extends javax.swing.JDialog {
                    tampilkan_ppnobat_ralan="";
     private String[] keranap,kodebarang,namabarang,kategori,satuan;
     private Double[] kapasitas,stok,harga,hargabeli,subtotal;
+    private boolean sukses=false;
 
     /** Creates new form DlgProgramStudi
      * @param parent
@@ -255,7 +256,7 @@ public class DlgInputStokPasien extends javax.swing.JDialog {
 
         ppBersihkan.setBackground(new java.awt.Color(255, 255, 254));
         ppBersihkan.setFont(new java.awt.Font("Tahoma", 0, 11)); // NOI18N
-        ppBersihkan.setForeground(new java.awt.Color(70, 70, 70));
+        ppBersihkan.setForeground(new java.awt.Color(50,50,50));
         ppBersihkan.setIcon(new javax.swing.ImageIcon(getClass().getResource("/picture/stop_f2.png"))); // NOI18N
         ppBersihkan.setText("Bersihkan Jumlah");
         ppBersihkan.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
@@ -281,7 +282,7 @@ public class DlgInputStokPasien extends javax.swing.JDialog {
         setUndecorated(true);
         setResizable(false);
 
-        internalFrame1.setBorder(javax.swing.BorderFactory.createTitledBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(240, 245, 235)), "::[ Stok Obat & BHP Medis Pasien Di Ranap ]::", javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Tahoma", 0, 11), new java.awt.Color(70, 70, 70))); // NOI18N
+        internalFrame1.setBorder(javax.swing.BorderFactory.createTitledBorder(javax.swing.BorderFactory.createLineBorder(new java.awt.Color(240, 245, 235)), "::[ Stok Obat & BHP Medis Pasien Di Ranap ]::", javax.swing.border.TitledBorder.DEFAULT_JUSTIFICATION, javax.swing.border.TitledBorder.DEFAULT_POSITION, new java.awt.Font("Tahoma", 0, 11), new java.awt.Color(50,50,50))); // NOI18N
         internalFrame1.setName("internalFrame1"); // NOI18N
         internalFrame1.setLayout(new java.awt.BorderLayout(1, 1));
 
@@ -611,11 +612,12 @@ private void KdKeyPressed(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_TKdKey
             }else{
                 i= JOptionPane.showConfirmDialog(rootPane,"Eeiiiiiits, udah bener belum data yang mau disimpan..??","Konfirmasi",JOptionPane.YES_NO_OPTION);
                 if (i == JOptionPane.YES_OPTION) {
-                    
+                    Sequel.AutoComitFalse();
+                    sukses=true;
                     ttlhpp=0;ttljual=0;
                     for(i=0;i<tbDokter.getRowCount();i++){  
                        if(Valid.SetAngka(tbDokter.getValueAt(i,0).toString())>0){
-                            if(Sequel.menyimpantf("stok_obat_pasien","?,?,?,?,?","Stok Obat Pasien",5,new String[]{                            
+                            if(Sequel.menyimpantf2("stok_obat_pasien","?,?,?,?,?","Stok Obat Pasien",5,new String[]{                            
                                 Valid.SetTgl(Tgl.getSelectedItem()+""),norawat.getText(),tabMode.getValueAt(i,1).toString(),
                                 tabMode.getValueAt(i,0).toString(),kdgudang.getText()
                             })==true){
@@ -624,30 +626,46 @@ private void KdKeyPressed(java.awt.event.KeyEvent evt) {//GEN-FIRST:event_TKdKey
                                 Trackobat.catatRiwayat(tabMode.getValueAt(i,1).toString(),0,Valid.SetAngka(tabMode.getValueAt(i,0).toString()),"Stok Pasien Ranap",akses.getkode(),kdgudang.getText(),"Simpan");
                                 Sequel.menyimpan("gudangbarang","'"+tabMode.getValueAt(i,1).toString()+"','"+kdgudang.getText()+"','-"+tabMode.getValueAt(i,0).toString()+"'", 
                                                 "stok=stok-'"+tabMode.getValueAt(i,0).toString()+"'","kode_brng='"+tabMode.getValueAt(i,1).toString()+"' and kd_bangsal='"+kdgudang.getText()+"'");
+                            }else{
+                                sukses=false;
                             }                            
                        }
                     }  
 
-                    Sequel.queryu("delete from tampjurnal");    
-                    if(ttljual>0){
-                        Sequel.menyimpan("tampjurnal","'"+Suspen_Piutang_Obat_Ranap+"','Suspen Piutang Obat Ranap','"+ttljual+"','0'","Rekening");    
-                        Sequel.menyimpan("tampjurnal","'"+Obat_Ranap+"','Pendapatan Obat Rawat Inap','0','"+ttljual+"'","Rekening");                              
-                    }
-                    if(ttlhpp>0){
-                        Sequel.menyimpan("tampjurnal","'"+HPP_Obat_Rawat_Inap+"','HPP Persediaan Obat Rawat Inap','"+ttlhpp+"','0'","Rekening");    
-                        Sequel.menyimpan("tampjurnal","'"+Persediaan_Obat_Rawat_Inap+"','Persediaan Obat Rawat Inap','0','"+ttlhpp+"'","Rekening");                              
-                    }
-                    jur.simpanJurnal(norawat.getText(),Valid.SetTgl(Tgl.getSelectedItem()+""),"U","PEMBERIAN OBAT RAWAT INAP PASIEN, DIPOSTING OLEH "+akses.getkode());                                                
+                    if(sukses==true){
+                        Sequel.queryu("delete from tampjurnal");    
+                        if(ttljual>0){
+                            Sequel.menyimpan2("tampjurnal","'"+Suspen_Piutang_Obat_Ranap+"','Suspen Piutang Obat Ranap','"+ttljual+"','0'","Rekening");    
+                            Sequel.menyimpan2("tampjurnal","'"+Obat_Ranap+"','Pendapatan Obat Rawat Inap','0','"+ttljual+"'","Rekening");                              
+                        }
+                        if(ttlhpp>0){
+                            Sequel.menyimpan2("tampjurnal","'"+HPP_Obat_Rawat_Inap+"','HPP Persediaan Obat Rawat Inap','"+ttlhpp+"','0'","Rekening");    
+                            Sequel.menyimpan2("tampjurnal","'"+Persediaan_Obat_Rawat_Inap+"','Persediaan Obat Rawat Inap','0','"+ttlhpp+"'","Rekening");                              
+                        }
+                        sukses=jur.simpanJurnal(norawat.getText(),Valid.SetTgl(Tgl.getSelectedItem()+""),"U","PEMBERIAN OBAT RAWAT INAP PASIEN, DIPOSTING OLEH "+akses.getkode());  
+                    }                                                 
 
+                    if(sukses==true){
+                        Sequel.Commit();
+                    }else{
+                        sukses=false;
+                        JOptionPane.showMessageDialog(null,"Terjadi kesalahan saat pemrosesan data, transaksi dibatalkan.\nPeriksa kembali data sebelum melanjutkan menyimpan..!!");
+                        Sequel.RollBack();
+                    }   
+                    Sequel.AutoComitTrue();
                     
-                    for(index=0;index<tbDokter.getRowCount();index++){   
-                        tbDokter.setValueAt("",index,0);        
+                    if(sukses==true){
+                        for(index=0;index<tbDokter.getRowCount();index++){   
+                            tbDokter.setValueAt("",index,0); 
+                            tbDokter.setValueAt(0,index,9);  
+                        }
+                        ttl=0;
+                        ttlhpp=0;
+                        ttljual=0;
+                        LTotal.setText("0");
+                        LPpn.setText("0");
+                        LTotalTagihan.setText("0");
                     }
-                    tampil();
-
-                    LTotal.setText("0");
-                    LPpn.setText("0");
-                    LTotalTagihan.setText("0");
                 }
             }
         }
@@ -1073,46 +1091,50 @@ private void BtnGudangActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIR
     private void getData() {
         int row=tbDokter.getSelectedRow();
         if(nmgudang.getText().trim().equals("")){
-            Valid.textKosong(kdgudang,"Asal Stok");
+             Valid.textKosong(kdgudang,"Asal Stok");
         }else if(row!= -1){         
-            int kolom=tbDokter.getSelectedColumn();  
-            if((kolom==1)||(kolom==0)){    
+             int kolom=tbDokter.getSelectedColumn();  
+             if((kolom==1)||(kolom==0)){    
                if(!tabMode.getValueAt(row,0).toString().equals("")){
-                   if(Double.parseDouble(tabMode.getValueAt(row,0).toString())>Double.parseDouble(tabMode.getValueAt(row,6).toString())){
-                        JOptionPane.showMessageDialog(null,"Maaf, Stok tidak cukup....!!!");
-                        TCari.requestFocus();
-                        tabMode.setValueAt("", row,0);  
-                   }else{
-                        y=Double.parseDouble(tabMode.getValueAt(tbDokter.getSelectedRow(),0).toString())*
-                            Double.parseDouble(tabMode.getValueAt(tbDokter.getSelectedRow(),7).toString());
-                            tbDokter.setValueAt(y,tbDokter.getSelectedRow(),9);
-                            
-                        ttl=0;
-                        y=0;
-                        int row2=tabMode.getRowCount();
-                        for(int r=0;r<row2;r++){ 
-                            try {
-                                if(Double.parseDouble(tabMode.getValueAt(r,0).toString())>0){
-                                    y=Double.parseDouble(tabMode.getValueAt(r,0).toString())*
-                                      Double.parseDouble(tabMode.getValueAt(r,7).toString());
-                                }                                 
-                            } catch (Exception e) {
-                                y=0;
-                            }
-                            ttl=ttl+y;
-                        }
-                        LTotal.setText(Valid.SetAngka(ttl));
-                        ppnobat=0;
-                        if(tampilkan_ppnobat_ralan.equals("Yes")){
-                            ppnobat=ttl*0.1;
-                            ttl=ttl+ppnobat;
-                            LPpn.setText(Valid.SetAngka(ppnobat));
-                            LTotalTagihan.setText(Valid.SetAngka(ttl));
-                        }
-                        TCari.setText("");
-                    }                                    
+                   try {
+                       if(Double.parseDouble(tabMode.getValueAt(row,0).toString())>Double.parseDouble(tabMode.getValueAt(row,6).toString())){
+                             JOptionPane.showMessageDialog(null,"Maaf, Stok tidak cukup....!!!");
+                             TCari.requestFocus();
+                             tabMode.setValueAt("", row,0);  
+                        }else{
+                             y=Double.parseDouble(tabMode.getValueAt(tbDokter.getSelectedRow(),0).toString())*
+                                 Double.parseDouble(tabMode.getValueAt(tbDokter.getSelectedRow(),7).toString());
+                             tbDokter.setValueAt(y,tbDokter.getSelectedRow(),9);
+                             TCari.setText("");
+                        } 
+                   } catch (Exception e) {
+                       tabMode.setValueAt("", row,0);
+                       tabMode.setValueAt(0, row,9);
+                   }                                       
+                }else{
+                    tabMode.setValueAt("", row,0);
+                    tabMode.setValueAt(0, row,9);
                 }                 
-            }
+             }
+             ttl=0;
+             for(int r=0;r<tabMode.getRowCount();r++){ 
+                 y=0;
+                 try {
+                     y=Double.parseDouble(tabMode.getValueAt(r,0).toString())*
+                       Double.parseDouble(tabMode.getValueAt(r,7).toString());
+                 } catch (Exception e) {
+                     y=0;
+                 }
+                 ttl=ttl+y;
+             }
+             LTotal.setText(Valid.SetAngka(ttl));
+             ppnobat=0;
+             if(tampilkan_ppnobat_ralan.equals("Yes")){
+                 ppnobat=ttl*0.1;
+                 ttl=ttl+ppnobat;
+                 LPpn.setText(Valid.SetAngka(ppnobat));
+                 LTotalTagihan.setText(Valid.SetAngka(ttl));
+             }
         }
     }
